requirements.txt
flask-restful
black==21.6b0
requests==2.25.1
pytest==6.2.4
Faker==8.8.2
pymongo==3.11.4
mongoengine==0.23.0